feat(书签): 添加书签删除功能和相关UI操作
实现书签删除功能，包括：
1. 在BookmarkUseCases、BookmarkRepository和ReadeckApiClient中添加删除方法
2. 在BookmarkDetailViewModel中添加删除命令和操作逻辑
3. 在书签详情页添加删除按钮和确认对话框
4. 将标签编辑对话框提取为独立组件以便复用
